Move profile card to account settings
